backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
python-multipart
passlib[bcrypt]
python-jose[cryptography]
cors
